🚚conflict 없는 master 머지를 위한 pages 경로 수정
