+++ src/index.jsx
@@ -2,8 +2,7 @@ import { ChakraProvider } from '@chakra-ui/core';
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
-import MainPage from './pages/index';
-import CategoryPage from './pages/CategoryPage';
+import { MainPage, CategoryPage } from './pages';
 import theme from './themes/index';
 
 ReactDOM.render(
