+++ src/pages/index.js
@@ -1,15 +1,2 @@
-import { Heading, Box } from '@chakra-ui/core';
-import React from 'react';
-import SearchBar from '../components/SearchBar';
-import Layout from '../components/Layout';
-
-const MainPage = () => (
-  <Layout>
-    <Heading color="main">Avocado</Heading>
-    <Box mt="76px" w="full">
-      <SearchBar mt="76px" />
-    </Box>
-  </Layout>
-);
-
-export default MainPage;
+export { default as MainPage } from './MainPage';
+export { default as CategoryPage } from './CategoryPage';
